Reorganized radio
