+++ src/space/core/chat/mod.rs
@@ -1,3 +1,4 @@
 pub mod events;
 pub mod functions;
 pub mod systems;
+pub mod components;
